+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 
-## Version 1.0.3 (2016-05-20)
+## Version 1.0.5 (COMING)
+* StatusBarUtils.setLightStatusBar(View _view) added to apply a light status bar (>= Marshmallow)
+* StatusBarUtils.isLight(int _color) added, allows to know if a given color is light or not
+
+## Version 1.0.4 (2016-05-20)
 * SquareImageView added for easier ImageView's inside GridViews
 
 ## Version 1.0.2 (2016-05-18)